🔨 refactor: ProfileImg 컴포넌트 공통 컴포넌트로 이동
