+++ src/components/Mypage/Sidebar/index.tsx
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from 'react-router-dom';
-import ProfileImg from '../ProfileImg';
+import ProfileImg from '../../Common/ProfileImg';
 import SidebarItem from '../SidebarItem';
 import { Container, GoButton, ProfileContainer } from './index.style';
